Logics of meme getting were added in presenter, wrong type of MemesDto fields were fixed

+++ app/src/main/java/com/meme/model/dto/MemeDto.kt
@@ -4,15 +4,15 @@ import com.google.gson.annotations.SerializedName
 
 data class MemeDto(
     @SerializedName("id")
-    var id: Int,
+    var id: Long,
     @SerializedName("title")
-    var title:String,
+    var title: String,
     @SerializedName("description")
     var description: String,
     @SerializedName("isFavourite")
     var isFavourite: Boolean,
     @SerializedName("createdDate")
-    var createdDate: Int,
+    var createdDate: Long,
     @SerializedName("photoUrl")
     var photoUrl: String
 )
